+++ cpp/src/barretenberg/honk/composer/composer_helper/standard_honk_composer_helper.hpp
@@ -10,8 +10,8 @@
 #include "barretenberg/proof_system/verification_key/verification_key.hpp"
 #include "barretenberg/plonk/proof_system/verifier/verifier.hpp"
 #include "barretenberg/proof_system/composer/composer_base.hpp"
-#include "composer_helper_lib.hpp"
-#include "permutation_helper.hpp"
+#include "barretenberg/proof_system/composer/composer_helper_lib.hpp"
+#include "barretenberg/proof_system/composer/permutation_helper.hpp"
 
 #include <utility>
