+++ cpp/src/barretenberg/proof_system/composer/composer_helper_lib.hpp
