+++ cpp/src/barretenberg/proof_system/composer/permutation_helper.hpp
